fix(clerk-js): Move getFullName and getIdentifier to core/utils
To enable code splitting of the /ui dir, /core should never depend on it. It should always be the other way around.
