+++ packages/clerk-js/src/ui/components/ImpersonationFab/ImpersonationFab.tsx
@@ -1,6 +1,7 @@
 import type { PointerEventHandler } from 'react';
 import React, { useEffect, useRef } from 'react';
 
+import { getFullName, getIdentifier } from '../../../utils/user';
 import { useCoreClerk, useCoreSession, withCoreUserGuard } from '../../contexts';
 import type { LocalizationKey } from '../../customizables';
 import {
@@ -18,7 +19,6 @@ import { Portal } from '../../elements/Portal';
 import { Eye } from '../../icons';
 import type { PropsOfComponent } from '../../styledSystem';
 import { mqu } from '../../styledSystem';
-import { getFullName, getIdentifier } from '../../utils';
 
 type EyeCircleProps = PropsOfComponent<typeof Col> & {
   width: string;
